rename settings module to commons
