+++ app/modules/commons/models.py
@@ -1,14 +1,12 @@
-# -*- coding: utf-8 -*-
+"""Models for the commons module."""
 
-# Getting config data
 from config import tz
-
-# Import the database object (db) from the main application module
 from app import db
 
 
-# Function to format an object (like datetime/date) to a string
 def default_object_string(object, timezone=tz):
+    """Function to format an object (like datetime/date) to a string."""
+
     if str(type(object)) == "<class 'datetime.datetime'>":
         try:
             return (
@@ -21,8 +19,9 @@ def default_object_string(object, timezone=tz):
     return object
 
 
-# Define a base model for other database tables to inherit
 class Base(db.Model):
+    """Base application model for other database tables to inherit."""
+
     __abstract__ = True
 
     # Defining base columns
@@ -35,7 +34,6 @@ class Base(db.Model):
     )
 
 
-# Define a UF model using Base columns
 class UF(Base):
     __tablename__ = "uf"
 
@@ -47,7 +45,6 @@ class UF(Base):
     city = db.relationship("City", lazy="select", backref="uf")
     # model_name = db.relationship('ModelName', lazy='select', backref='uf')
 
-    # New instance instantiation procedure
     def __init__(self, code, name):
         self.code = code
         self.name = name
@@ -57,19 +54,17 @@ class UF(Base):
 
     # Returning data as dict
     def as_dict(self, timezone=tz):
-        # We also remove the password
         data = {
             c.name: default_object_string(getattr(self, c.name), timezone)
             for c in self.__table__.columns
         }
-        # Adding the related tables
+        # Add the related tables
         for c in self.__dict__:
             if "app" in str(type(self.__dict__[c])):
                 data[c] = self.__dict__[c].as_dict(timezone)
         return data
 
 
-# Define a city model using Base columns
 class City(Base):
     __tablename__ = "city"
 
@@ -82,7 +77,6 @@ class City(Base):
     # Relationships
     # model_name = db.relationship('ModelName', lazy='select', backref='city')
 
-    # New instance instantiation procedure
     def __init__(self, name, uf_id):
         self.name = name
         self.uf_id = uf_id
@@ -92,12 +86,11 @@ class City(Base):
 
     # Returning data as dict
     def as_dict(self, timezone=tz):
-        # We also remove the password
         data = {
             c.name: default_object_string(getattr(self, c.name), timezone)
             for c in self.__table__.columns
         }
-        # Adding the related tables
+        # Add the related tables
         for c in self.__dict__:
             if "app" in str(type(self.__dict__[c])):
                 data[c] = self.__dict__[c].as_dict(timezone)
