+++ app/modules/commons/swagger/city/delete_item.yml
