+++ app/modules/commons/swagger/city/get_item_by_id.yml
