+++ app/modules/commons/swagger/city/index_item.yml
