+++ app/modules/commons/swagger/uf/delete_item.yml
